split out request parameter methods
