+++ src/utils/common.js
@@ -21,89 +21,6 @@ function hexToStr(hexString) {
   return str;
 }
 
-function getInt(s) {
-  return new Promise(function(resolve, reject) {
-    if (_.isNumber(s) && s >> 0 === s) {
-      return resolve(s);
-    }
-    let x = parseInt(s);
-    if (isNaN(x)) {
-      return reject("polis_fail_parse_int " + s);
-    }
-    resolve(x);
-  });
-}
-
-function getBool(s) {
-  return new Promise(function(resolve, reject) {
-    let type = typeof s;
-    if ("boolean" === type) {
-      return resolve(s);
-    }
-    if ("number" === type) {
-      if (s === 0) {
-        return resolve(false);
-      }
-      return resolve(true);
-    }
-    s = s.toLowerCase();
-    if (s === 't' || s === 'true' || s === 'on' || s === '1') {
-      return resolve(true);
-    } else if (s === 'f' || s === 'false' || s === 'off' || s === '0') {
-      return resolve(false);
-    }
-    reject("polis_fail_parse_boolean");
-  });
-}
-
-function getIntInRange(min, max) {
-  return function(s) {
-    return getInt(s).then(function(x) {
-      if (x < min || max < x) {
-        throw "polis_fail_parse_int_out_of_range";
-      }
-      return x;
-    });
-  };
-}
-
-function extractFromQueryOrBody(req, name) {
-  let ret = extractFromQuery(req, name);
-  if (ret) {
-    return ret;
-  } else {
-    return extractFromBody(req, name);
-  }
-}
-
-function extractFromQuery(req, name) {
-  if (name in req.query) {
-    return req.query[name];
-  }
-  return void 0;
-}
-
-function extractFromBody(req, name) {
-  if (!req.body) {
-    return void 0;
-  }
-  return req.body[name];
-}
-
-function extractFromCookie(req, name) {
-  if (!req.cookies) {
-    return void 0;
-  }
-  return req.cookies[name];
-}
-
-function extractFromHeader(req, name) {
-  if (!req.headers) {
-    return void 0;
-  }
-  return req.headers[name.toLowerCase()];
-}
-
 const polisTypes = {
   reactions: {
     push: 1,
@@ -127,13 +44,5 @@ polisTypes.starValues = _.values(polisTypes.staractions);
 module.exports = {
   strToHex,
   hexToStr,
-  getInt,
-  getBool,
-  getIntInRange,
-  extractFromQueryOrBody,
-  extractFromQuery,
-  extractFromBody,
-  extractFromCookie,
-  extractFromHeader,
   polisTypes
 };
